fix: Rename .privateDemo.js to demoFragment.js
